Conclusão Services

+++ service/beneficios_service.py
@@ -4,14 +4,27 @@ class BeneficiosService:
     def __init__(self, repository = None):
         self.repository = repository or BeneficiosRepository()
 
-    def create_beneficio(self, nome):
+    def create(self, nome):
+
+        if self.repository.get_by_nome(nome):
+            raise ValueError("Benefício já existente.") 
+        
         return self.repository.create(nome)
 
-    def get_beneficio_by_id(self, id_beneficios):
-        return self.repository.get_by_id(id_beneficios)
+    def get_by_id(self, id_beneficios):
 
-    def get_all_beneficios(self):
+        beneficio = self.repository.get_by_id(id_beneficios)
+        if not beneficio:   
+            raise ValueError("Benefício não encontrado.")  
+         
+        return beneficio
+
+    def get_all(self):
         return self.repository.get_all()
     
-    def delete_beneficio(self, id_beneficios):
+    def delete(self, id_beneficios):
+
+        beneficio = self.repository.get_by_id(id_beneficios)
+        if not beneficio:   
+            raise ValueError("Benefício não encontrado.")
         return self.repository.delete(id_beneficios)
